[ticket#74] Nun kann man sich auch per Hilfe die aktuelle KeyMap ansehen.

+++ structs.h
@@ -341,7 +341,7 @@ struct POINT_STRUCT
 struct C64_KEYS
 {
     unsigned char MatrixCode;
-    unsigned long SDLKeyCode;
+    int SDLKeyCode;
     bool          Shift;
 };
